+++ src/utils/index.ts
@@ -1,5 +1,6 @@
 export { buildSchema, buildSchemaSync, BuildSchemaOptions } from "./buildSchema";
 export { buildTypeDefsAndResolvers } from "./buildTypeDefsAndResolvers";
+export { createResolversMap } from "./createResolversMap";
 export {
   emitSchemaDefinitionFile,
   emitSchemaDefinitionFileSync,